# Conflicts:
#   back/src/main/java/gg/garen/back/matchPrediction/dto/ParticipantDto.java

Former-commit-id: 43ca6fb8c3f7bc17087d8533d8a7a5e48de288d9

+++ back/src/main/java/gg/garen/back/common/domain/mysql/PlayerInfo.java
@@ -26,13 +26,13 @@ public class PlayerInfo extends Auditable {
     int leaguePoints;
     int wins;
     int losses;
-    int apiKeyId;
+    Long apiKeyId;
 
-    public int getApiKeyId() {
+    public Long getApiKeyId() {
         return apiKeyId;
     }
 
-    public void setApiKeyId(int apiKeyId) {
+    public void setApiKeyId(Long apiKeyId) {
         this.apiKeyId = apiKeyId;
     }
